pyproject.toml
[project]
name = "kmm-tools"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "graphviz>=0.20.3",
    "pandas>=2.0.3",
    "python-dotenv>=1.0.1",
    "streamlit>=1.40.1",
]
